refactor: Moved out filter module
