+++ cognite/client/data_classes/data_modeling/__init__.py
@@ -1,4 +1,5 @@
-from cognite.client.data_classes.data_modeling import aggregations, filters, query
+from cognite.client.data_classes import filters
+from cognite.client.data_classes.data_modeling import aggregations, query
 from cognite.client.data_classes.data_modeling.aggregations import AggregatedValue, Aggregation
 from cognite.client.data_classes.data_modeling.containers import (
     Constraint,
@@ -39,7 +40,6 @@ from cognite.client.data_classes.data_modeling.data_types import (
     TimeSeriesReference,
     Timestamp,
 )
-from cognite.client.data_classes.data_modeling.filters import Filter
 from cognite.client.data_classes.data_modeling.ids import (
     ContainerId,
     ContainerIdentifier,
@@ -84,6 +84,7 @@ from cognite.client.data_classes.data_modeling.views import (
     ViewFilter,
     ViewList,
 )
+from cognite.client.data_classes.filters import Filter
 
 # TODO: Remove these in next major version, there just here to ensure backwards compatibility after renaming and
 #  removing some data classes.
@@ -105,11 +106,11 @@ __all__ = [
     "DataModelingId",
     "ContainerIdentifier",
     "DataModelIdentifier",
-    "filters",
     "DirectRelation",
     "ViewDirectRelation",
     "ContainerDirectRelation",
     "Filter",
+    "filters",
     "DirectRelationReference",
     "DataModel",
     "DataModelList",
